Vendor-Application/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
supabase>=2.3.0
cryptography>=42.0.8
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
scikit-learn>=1.3.0
Pillow>=10.0.0
numpy>=1.26.0
python-multipart>=0.0.9